Refactor import statements to use correct paths for context files
